+++ src/main/java/_persistence/rubric/RubricsLoader.java
@@ -1,0 +1,55 @@
+package _persistence.rubric;
+
+import java.sql.Connection;
+import java.util.List;
+
+import _error.ImproperRubricInfo;
+import rubric.Rubric;
+
+/**
+ * The [RubricsLoader] class...
+ */
+public class RubricsLoader {
+  private RubricQueryBuilder queryBuilder;
+
+  /**
+   * The [RubricsLoader] constructor...
+   */
+  RubricsLoader (Connection connection) {
+    queryBuilder = new RubricQueryBuilder (connection);
+  }
+
+  /**
+   * The [loadRubricsByCourseId] method...
+   */
+  List<Rubric> loadRubricsByCourseId (String courseId) throws ImproperRubricInfo {
+    RubricQueryExecutor rubricExecutor;
+
+    try {
+      rubricExecutor = new RubricQueryExecutor (
+        queryBuilder.buildRetrieveRubricsByCourseIdQuery (courseId)
+      );
+    } catch (Exception e) {
+      throw new ImproperRubricInfo (e.getMessage());
+    }
+
+    return rubricExecutor.retrieveRubricsForCourse();
+  }
+
+  /**
+   * The [loadRubricById] method...
+   */
+  Rubric loadRubricById (String rubricId) throws ImproperRubricInfo {
+    RubricQueryExecutor rubricExecutor;
+
+    try {
+      rubricExecutor = new RubricQueryExecutor (
+          queryBuilder.buildRetrieveRubricByIdQuery (rubricId)
+      );
+    } catch (Exception e) {
+      throw new ImproperRubricInfo (e.getMessage());
+    }
+
+    return rubricExecutor.retrieveRubric();
+  }
+}